+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 ## Added
 - community: overseer task manager for makefile and other tasks, leader x tasks mappings
+- community: add spectre plugin for project wide search and replace
 
 # 2023-11-13
 ## Added